requirements.txt
Pillow~=9.1.0
mysql-connector-python~=8.0.28